move local storage module to src/utils/
